[MOD/#22] 뷰 이름 수정
